% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbs4_add_date_column.R
\name{cbs4_add_date_column}
\alias{cbs4_add_date_column}
\title{Convert the time variable into either a date or numeric.}
\usage{
cbs4_add_date_column(data, date_type = c("Date", "numeric"), ...)
}
\arguments{
\item{data}{\code{data.frame} retrieved using \code{\link[=cbs4_get_data]{cbs4_get_data()}}}

\item{date_type}{Type of date column: "Date", "numeric. Numeric creates a fractional
number which signs the "middle" of the period. e.g. 2018JJ00 -> 2018.5 and
2018KW01 -> 2018.167. This is for the following reasons: otherwise 2018.0 could mean
2018, 2018 Q1 or 2018 Jan, and furthermore 2018.75 is a bit strange for 2018 Q4.
If all codes in the dataset have frequency "Y" the numeric output will be \code{integer}.}

\item{...}{future use.}
}
\value{
original dataset with two added columns: \verb{<period>_date} and
\verb{<period>_freq}. This last column is a factor with levels: \code{Y}, \code{Q} and \code{M}
}
\description{
Time periods in data of CBS are coded: yyyyXXww (e.g. 2018JJ00, 2018MM10, 2018KW02),
which contains year (yyyy), type (XX) and index (ww). \code{cbs4_add_date_column} converts
these codes into a \code{\link[=Date]{Date()}} or \code{numeric}. In addition it adds
a frequency column denoting the type of the column.
}
\examples{
if (interactive()){
  # works on observations...
  obs <- cbs4_get_observations( id        = "80784ned"    # table id
                              , Perioden  = "2019JJ00" # Year 2019
                              , Geslacht  = "1100"       # code for total gender
                              , RegioS    = "NL01"       # code for region NL
                              , Measure  = "M003371_2"
                              )

  # add a Periods_Date column
  obs_d <- cbs4_add_date_column(obs)
  obs_d

  # add a Periods_numeric column
  obs_d <- cbs4_add_date_column(obs, date_type = "numeric")
  obs_d

  # works on data
  d <- cbs4_get_data( id        = "80784ned"    # table id
                    , Perioden  = "2019JJ00"   # Year 2019
                    , Geslacht  = "1100"       # code for total gender
                    , RegioS    = "NL01"       # code for region NL
                    , Measure   = "M003371_2"
                    )
  cbs4_add_date_column(d)
}
}
\seealso{
Other add metadata: 
\code{\link{cbs4_add_labels}()}
}
\concept{add metadata}
