% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbs4_add_label_columns.R
\name{cbs4_add_label_columns}
\alias{cbs4_add_label_columns}
\title{Add understandable labels to a table}
\usage{
cbs4_add_label_columns(data, ...)
}
\arguments{
\item{data}{downloaded with \code{\link[=cbs4_get_data]{cbs4_get_data()}}}

\item{...}{not used}
}
\value{
original dataset with extra label columns. See details.
}
\description{
Add columns with labels to the dataset.
}
\details{
\code{\link[=cbs4_add_label_columns]{cbs4_add_label_columns()}} adds for the \code{Measure} and each \verb{<Dimension>}
column an extra column \code{MeasureLabel} ( \verb{<Dimension>Label}) that contains
the \code{Title} of each code, making the table more digestible. \code{Title} and
other metadata can also be found using \code{\link[=cbs4_get_metadata]{cbs4_get_metadata()}}.
}
\examples{
if (interactive()){
  # works on observations
  obs <- cbs4_get_observations("84287NED", Perioden="2019MM12")
  obs # without label columns

  obs_labeled <- cbs4_add_label_columns(obs)
  obs_labeled

  # works on data
  d <- cbs4_get_data("84287NED", Perioden="2019MM12")
  d # cbs4_get_data automagically labels measure columns.

  d_labeled <- cbs4_add_label_columns(d)
  d_labeled
}
}
\seealso{
\code{\link[=cbs4_get_metadata]{cbs4_get_metadata()}}

Other add metadata columns: 
\code{\link{cbs4_add_date_column}()},
\code{\link{cbs4_add_unit_column}()}
}
\concept{add metadata columns}
