% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbs4_add_unit.R
\name{cbs4_add_unit}
\alias{cbs4_add_unit}
\title{Add unit column to data}
\usage{
cbs4_add_unit(data, ...)
}
\arguments{
\item{data}{downloaded with \code{\link[=cbs4_get_observations]{cbs4_get_observations()}}}

\item{...}{not used}
}
\description{
Add labels to codes
}
\examples{
if (interactive()){
  # works only on observations
  obs <- cbs4_get_observations("84287NED", Perioden="2019MM12")
  obs # without Unit column

  obs_unit <- cbs4_add_unit(obs)
  obs_unit # with unit column
}
}
