% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbs4_download.R
\name{cbs4_download}
\alias{cbs4_download}
\title{Download observations and metadata}
\usage{
cbs4_download(
  id,
  download_dir = id,
  ...,
  catalog = "CBS",
  show_progress = interactive(),
  sep = ",",
  verbose = FALSE
)
}
\arguments{
\item{id}{Identifier of publication}

\item{download_dir}{directory where files are to be stored}

\item{...}{optional selection statement to retrieve a subset of the data.}

\item{catalog}{Catalog to download from}

\item{show_progress}{\code{logical} if \code{TRUE} downloading shows a progress bar. Cannot be used
together with \code{verbose=TRUE}}

\item{sep}{seperator to be used in writing the data}

\item{verbose}{Should messages be printed...}
}
\description{
Download observations and metadata to a directory
}
\seealso{
Other data-download: 
\code{\link{cbs4_get_data}()},
\code{\link{cbs4_get_observations}()}
}
\concept{data-download}
