% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbs4_get_datasets.R
\name{cbs4_get_datasets}
\alias{cbs4_get_datasets}
\alias{cbs4_get_toc}
\title{Get available datasets}
\usage{
cbs4_get_datasets(
  catalog = "CBS",
  convert_dates = TRUE,
  base_url = BASEURL4,
  verbose = getOption("cbsodata4.verbose", FALSE)
)

cbs4_get_toc(
  catalog = "CBS",
  convert_dates = TRUE,
  base_url = BASEURL4,
  verbose = getOption("cbsodata4.verbose", FALSE)
)
}
\arguments{
\item{catalog}{only show the datasets from that catalog. If \code{NULL}
all datasets of all catalogs will be returned.}

\item{convert_dates}{Converts date columns in Date-Time type (in stead of \code{character})}

\item{base_url}{base url of the CBS OData 4 API}

\item{verbose}{Should the url request be printed?}
}
\description{
Get the available datasets from open data portal statline.
}
\details{
Setting the catalog to \code{NULL} will return all
}
\note{
the datasets are downloaded only once per R session and cached. Subsequent calls to
\code{cbs4_get_datasets} will use the results of the first call.
}
\seealso{
Other datasets: 
\code{\link{cbs4_get_catalogs}()}
}
\concept{datasets}
