% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbs4_get_metadata.R
\name{cbs4_get_metadata}
\alias{cbs4_get_metadata}
\title{Retrieve the metadata of a publication}
\usage{
cbs4_get_metadata(
  id,
  catalog = "CBS",
  ...,
  base_url = BASEURL4,
  verbose = getOption("cbsodataR.verbose", FALSE)
)
}
\arguments{
\item{id}{Identifier of publication or data retrieved with \code{\link[=cbs4_get_data]{cbs4_get_data()}}}

\item{catalog}{Catalog, @seealso \code{\link[=cbs4_get_catalogs]{cbs4_get_catalogs()}}}

\item{...}{not used}

\item{base_url}{alternative url that implements same interface as statistics netherlands.}

\item{verbose}{Should the function report on retrieving the data}
}
\description{
Retrieve the metadata of a publication
}
