% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbs4_get_data.R
\name{cbs4_get_observations}
\alias{cbs4_get_observations}
\title{Get observations from cbs odata}
\usage{
cbs4_get_observations(
  id,
  ...,
  catalog = "CBS",
  download_dir = file.path(tempdir(), id),
  verbose = FALSE,
  show_progress = interactive(),
  sep = ",",
  includeId = TRUE,
  as.data.table = FALSE
)
}
\arguments{
\item{id}{Identifier of the Opendata table. Can be retrieved with \code{\link[=cbs4_get_datasets]{cbs4_get_datasets()}}}

\item{...}{passed through to cbs4_download_data}

\item{catalog}{Catalog in which the dataset is to be found}

\item{download_dir}{directory in which the data and metadata is downloaded. By default this is
temporary directory, but can be set manually}

\item{verbose}{if \code{TRUE} prints the steps taken to retrieve the data.}

\item{show_progress}{if \code{TRUE} shows progress of data donwload, can't be used together
with verbose.}

\item{sep}{separator to be used to download the data.}

\item{includeId}{\code{logical}, should the Id column be downloaded?}

\item{as.data.table}{\code{logical}, should the result be of type data.table?}
}
\description{
Get data from the content cijfer bank
}
