% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_labels.R
\name{ccb_add_labels}
\alias{ccb_add_labels}
\title{Add labels to codes}
\usage{
ccb_add_labels(data, ...)
}
\arguments{
\item{data}{downloaded with \code{\link{ccb_get_data}}}

\item{...}{not used}
}
\description{
Add labels to codes
}
