% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_unit.R
\name{ccb_add_unit}
\alias{ccb_add_unit}
\title{Add unit column to data}
\usage{
ccb_add_unit(data, ...)
}
\arguments{
\item{data}{downloaded with \code{\link{ccb_get_data}}}

\item{...}{not used}
}
\description{
Add labels to codes
}
