% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{ccb_download_data}
\alias{ccb_download_data}
\title{Download the data file}
\usage{
ccb_download_data(id, download_dir = id, catalog = "CBS", ...,
  sep = ",", verbose = TRUE)
}
\arguments{
\item{id}{Identifier of publication}

\item{download_dir}{directory where files are to be stored}

\item{catalog}{Catalog to download from}

\item{...}{not used}

\item{verbose}{Should messages be printed...}
}
\description{
Download the data file
}
