% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{ccb_get_metadata}
\alias{ccb_get_metadata}
\title{Retrieve the metadata of a publication}
\usage{
ccb_get_metadata(id, catalog = "CBS", ..., verbose = FALSE)
}
\arguments{
\item{id}{Identifier of publication}

\item{catalog}{Catalog, @seealso \code{\link{ccb_get_catalogs}}}

\item{...}{not used}

\item{verbose}{Should the function report on retrieving the data}
}
\description{
Retrieve the metadata of a publication
}
