% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbs4_add_labels.R
\name{cbs4_add_labels}
\alias{cbs4_add_labels}
\title{Add labels to codes}
\usage{
cbs4_add_labels(data, ...)
}
\arguments{
\item{data}{downloaded with \code{\link[=cbs4_get_data]{cbs4_get_data()}}}

\item{...}{not used}
}
\description{
Add labels to codes
}
\seealso{
Other add metadata: 
\code{\link{cbs4_add_date_column}()}
}
\concept{add metadata}
