% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbs4_download.R
\name{cbs4_download}
\alias{cbs4_download}
\title{Download observations and metadata}
\usage{
cbs4_download(
  id,
  download_dir = id,
  ...,
  query = NULL,
  catalog = "CBS",
  show_progress = interactive() && !verbose,
  sep = ",",
  verbose = getOption("cbsodata4.verbose", FALSE),
  base_url = getOption("cbsodata4.base_url", BASEURL4)
)
}
\arguments{
\item{id}{Identifier of publication}

\item{download_dir}{directory where files are to be stored}

\item{...}{optional selection statement to retrieve a subset of the data.}

\item{query}{optional odata4 query in odata syntax (overwrites any specification in \code{...})}

\item{catalog}{Catalog to download from}

\item{show_progress}{\code{logical} if \code{TRUE} downloading shows a progress bar. Cannot be used
together with \code{verbose=TRUE}}

\item{sep}{seperator to be used in writing the data}

\item{verbose}{Should messages be printed...}

\item{base_url}{Possible other website which implements same protocol.}
}
\description{
Download observations and metadata to a directory. This function is the working
horse for \code{\link[=cbs4_get_data]{cbs4_get_data()}} and \code{\link[=cbs4_get_observations]{cbs4_get_observations()}} and has many of the same
options. This function is useful if you do not want to load an entire dataset
into memory, but just download the data and metadata in csv format.
}
\seealso{
Other data-download: 
\code{\link{cbs4_get_data}()},
\code{\link{cbs4_get_observations}()}
}
\concept{data-download}
