% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbs4_get_catalogs.R
\name{cbs4_get_catalogs}
\alias{cbs4_get_catalogs}
\title{Retrieve all (alternative) catalogs of Statistics Netherlands}
\usage{
cbs4_get_catalogs(
  base_url = getOption("cbsodata4.base_url", BASEURL4),
  verbose = getOption("cbsodata4.verbose", FALSE)
)
}
\arguments{
\item{base_url}{possible other url that implements same interface}

\item{verbose}{if \code{TRUE} the communication to the server is shown.}
}
\description{
Retrieve catalogs of Statistics Netherlands. Beside the main "CBS" catalog
other catalogs contain extra datasets that are not part of the main production
of SN.
}
\seealso{
Other datasets: 
\code{\link{cbs4_get_datasets}()}
}
\concept{datasets}
