% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_catalogs.R
\name{cbs4_get_catalogs}
\alias{cbs4_get_catalogs}
\title{Retrieve catalogs from ccb}
\usage{
cbs4_get_catalogs(
  base_url = BASEURL4,
  verbose = getOption("cbsodataR.verbose", FALSE)
)
}
\arguments{
\item{base_url}{Default}

\item{verbose}{if \code{TRUE} the communication to the server is shown.}
}
\description{
Retrieve catalogs from ccb
}
\seealso{
Other datasets: 
\code{\link{cbs4_get_datasets}()}
}
\concept{datasets}
