% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbs4_get_metadata.R
\name{cbs4_get_metadata}
\alias{cbs4_get_metadata}
\title{Retrieve the metadata of a publication}
\usage{
cbs4_get_metadata(
  id,
  catalog = "CBS",
  ...,
  base_url = BASEURL4,
  verbose = getOption("cbsodata4.verbose", FALSE)
)
}
\arguments{
\item{id}{Identifier of publication or data retrieved with \code{\link[=cbs4_get_data]{cbs4_get_data()}}/\code{\link[=cbs4_get_observations]{cbs4_get_observations()}}}

\item{catalog}{Catalog, from the set of \code{\link[=cbs4_get_catalogs]{cbs4_get_catalogs()}}}

\item{...}{not used}

\item{base_url}{alternative url that implements same interface as statistics netherlands.}

\item{verbose}{Should the function report on retrieving the data}
}
\description{
Retrieve the metadata of a publication.
The meta object contains all metadata properties of cbsodata in the form
of data.frames. Each \code{data.frame} describes properties of the SN
table: “Dimensions”, “MeasureCodes” and one ore more
“\<Dimension\>Codes” describing the meta data of the borders of a SN
table.
}
