% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{cbs4_get_metadata}
\alias{cbs4_get_metadata}
\title{Retrieve the metadata of a publication}
\usage{
cbs4_get_metadata(id, catalog = "CBS", ..., verbose = FALSE)
}
\arguments{
\item{id}{Identifier of publication or data retrieved with \code{\link{cbs4_get_data}}}

\item{catalog}{Catalog, @seealso \code{\link{cbs4_get_catalogs}}}

\item{...}{not used}

\item{verbose}{Should the function report on retrieving the data}
}
\description{
Retrieve the metadata of a publication
}
