% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbs4_get_observations.R
\name{cbs4_get_observations}
\alias{cbs4_get_observations}
\title{Get observations from a table.}
\usage{
cbs4_get_observations(
  id,
  ...,
  catalog = "CBS",
  download_dir = file.path(tempdir(), id),
  verbose = FALSE,
  show_progress = interactive() && !verbose,
  sep = ",",
  includeId = TRUE,
  as.data.table = FALSE
)
}
\arguments{
\item{id}{Identifier of the Opendata table. Can be retrieved with \code{\link[=cbs4_get_datasets]{cbs4_get_datasets()}}}

\item{...}{optional selections on data, passed through to cbs4_download. See examples}

\item{catalog}{Catalog in which the dataset is to be found.}

\item{download_dir}{directory in which the data and metadata is downloaded. By default this is
temporary directory, but can be set manually}

\item{verbose}{if \code{TRUE} prints the steps taken to retrieve the data.}

\item{show_progress}{if \code{TRUE} shows progress of data donwload, can't be used together
with verbose.}

\item{sep}{separator to be used to download the data.}

\item{includeId}{\code{logical}, should the Id column be downloaded?}

\item{as.data.table}{\code{logical}, should the result be of type data.table?}
}
\description{
Get observations from table \code{id}. Observations are data of a CBS opendata table in so-called long
format. There is one \code{Measure} column denoting the different variables/topics, and there is a column
for each of the dimensions, which can be queried using \code{\link[=cbs4_get_metadata]{cbs4_get_metadata()}}. Furtheremore it
contains one \code{Value} columns contains the data. \code{\link[=cbs4_get_data]{cbs4_get_data()}} offers an alternative in which
each variable/topic/Measure has its own column.
}
\examples{
if (interactive()){

  # filter on Perioden (see meta$PeriodenCodes)
  cbs4_get_data("84287NED"
               , Perioden = "2019MM12" # december 2019
               )

  # filter on multiple Perioden (see meta$PeriodenCodes)
  cbs4_get_data("84287NED"
               , Perioden = c("2019MM12", "2020MM01") # december 2019, january 2020
               )

  # to filter on a dimension just add the filter to the query

  # filter on Perioden (see meta$PeriodenCodes)
  cbs4_get_data("84287NED"
               , Perioden = "2019MM12" # december 2019
               , BedrijfstakkenBranchesSBI2008 = "T001081"
               )


  # filter on Perioden with contains
  cbs4_get_data("84287NED"
                , Perioden = contains("2020")
                , BedrijfstakkenBranchesSBI2008 = "T001081"
  )

  # filter on Perioden with multiple contains
  cbs4_get_data("84287NED"
                , Perioden = contains(c("2019MM1", "2020"))
                , BedrijfstakkenBranchesSBI2008 = "T001081"
  )

  # filter on Perioden with contains or = "2019MM12
  cbs4_get_data("84287NED"
                , Perioden = contains("2020") | "2019MM12"
                , BedrijfstakkenBranchesSBI2008 = "T001081"
  )

  # This all works on observations too
  cbs4_get_observations( id        = "80784ned"    # table id
                       , Perioden  = "2019JJ00"    # Year 2019
                       , Geslacht  = "1100"        # code for total gender
                       , RegioS    = contains("PV")
                       , Measure   = "M003371_2"
                       )

}
}
\seealso{
cbs4_get_metadata

Other data-download: 
\code{\link{cbs4_download}()},
\code{\link{cbs4_get_data}()}
}
\concept{data-download}
