% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{ccb_get_data}
\alias{ccb_get_data}
\title{Get data from the content cijfer bank}
\usage{
ccb_get_data(id, catalog = "CBS", ...,
  download_dir = file.path(tempdir(), id), verbose = FALSE,
  sep = ",")
}
\description{
Get data from the content cijfer bank
}
