% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasets.R
\name{ccb_get_datasets}
\alias{ccb_get_datasets}
\title{Get the available datasets}
\usage{
ccb_get_datasets(base_url = BASEURL)
}
\arguments{
\item{base_url}{base url of the CCB API}
}
\description{
Get the available datasets from ccb
}
